# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TrainFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /plans:
    get:
      operationId: listPlans
      tags: [plans]
      summary: List workout plans
      parameters:
        - name: studentId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkoutPlanDto"
    post:
      operationId: createPlan
      tags: [plans]
      summary: Create workout plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutPlanDto"
  /plans/{planId}:
    get:
      operationId: getPlan
      tags: [plans]
      summary: Get full plan with days and exercises
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Full plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutPlanFullDto"
    put:
      operationId: updatePlan
      tags: [plans]
      summary: Update plan metadata
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlanRequest"
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutPlanDto"
  /plans/{planId}/days:
    post:
      operationId: addPlanDay
      tags: [plans]
      summary: Add a day to a plan
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDayRequest"
      responses:
        "201":
          description: Created day
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutDayDto"
  /plans/{planId}/days/{dayId}/exercises:
    post:
      operationId: addDayExercise
      tags: [plans]
      summary: Add exercise to a plan day
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: string
        - name: dayId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanExerciseRequest"
      responses:
        "201":
          description: Created plan exercise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanExerciseDto"
...
components:
...
  schemas:
...
    ExerciseDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        muscleGroup:
          type: string
        equipment:
          type: string
        description:
          type: string
        isGlobal:
          type: boolean
        isCustom:
          type: boolean
      required: [id, name, muscleGroup, equipment]
...
    WorkoutPlanDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        trainerId:
          type: string
        studentId:
          type: string
        isPublished:
          type: boolean
      required: [id, name, trainerId]
    WorkoutDayDto:
      type: object
      properties:
        id:
          type: string
        planId:
          type: string
        dayName:
          type: string
        focus:
          type: string
        orderIndex:
          type: integer
      required: [id, planId, dayName, focus]
    PlanExerciseDto:
      type: object
      properties:
        id:
          type: string
        dayId:
          type: string
        exerciseId:
          type: string
        sets:
          type: integer
        reps:
          type: string
        restSeconds:
          type: integer
        notes:
          type: string
        orderIndex:
          type: integer
      required: [id, dayId, exerciseId, sets, reps, restSeconds]
    WorkoutPlanFullDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        trainerId:
          type: string
        studentId:
          type: string
        isPublished:
          type: boolean
        days:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutDayFullDto"
      required: [id, name, trainerId, days]
    WorkoutDayFullDto:
      type: object
      properties:
        id:
          type: string
        planId:
          type: string
        dayName:
          type: string
        focus:
          type: string
        orderIndex:
          type: integer
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/PlanExerciseFullDto"
      required: [id, planId, dayName, focus, exercises]
    PlanExerciseFullDto:
      type: object
      properties:
        id:
          type: string
        dayId:
          type: string
        exerciseId:
          type: string
        sets:
          type: integer
        reps:
          type: string
        restSeconds:
          type: integer
        notes:
          type: string
        orderIndex:
          type: integer
        exercise:
          $ref: "#/components/schemas/ExerciseDto"
      required: [id, dayId, exerciseId, sets, reps, restSeconds, exercise]
    CreatePlanRequest:
      type: object
      properties:
        name:
          type: string
        studentId:
          type: string
        isPublished:
          type: boolean
      required: [name]
    UpdatePlanRequest:
      type: object
      properties:
        name:
          type: string
        studentId:
          type: string
        isPublished:
          type: boolean
    CreateDayRequest:
      type: object
      properties:
        dayName:
          type: string
        focus:
          type: string
        orderIndex:
          type: integer
      required: [dayName, focus]
    CreatePlanExerciseRequest:
      type: object
      properties:
        exerciseId:
          type: string
        sets:
          type: integer
        reps:
          type: string
        restSeconds:
          type: integer
        notes:
          type: string
        orderIndex:
          type: integer
      required: [exerciseId, sets, reps, restSeconds]
...