# lib/api-zod/src/generated/types/createPlanRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrainFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePlanRequest {
  name: string;
  studentId?: string;
  isPublished?: boolean;
}